+++ app/src/main/java/by/geekbrains/appnotes/domain/NoteEntity.java
@@ -1,0 +1,41 @@
+package by.geekbrains.appnotes.domain;
+
+public class NoteEntity {
+
+    private String title;
+    private String description;
+    private String date;
+
+    public NoteEntity(String title, String description, String date) {
+        this.title = title;
+        this.description = description;
+        this.date = date;
+    }
+
+    public NoteEntity() {
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public void setDate(String date) {
+        this.date = date;
+    }
+}
